requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pymongo==4.5.0
beautifulsoup4==4.12.2
requests==2.31.0
asyncio==3.4.3
aiohttp==3.8.6
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
dnspython==2.4.2
certifi==2023.7.22
pytz